Major refactor.  Validate passes.
But validate passes because the tests are still wired to the old code in "prototype".
The idea here is to separate out Paventhan's (cool) prototype from the barebones
version of the compilation-oriented version that will be going in soon.

In the process, the modules should also be broke down at a slightly finer grain, and the
abstraction boundaries between areas of the code should be increased, with export
 lists reduced.